Backend/221095514-Distance-Sort
This is a utility script based on the user story that users would want to be able to find their
closest stores. It that can be integrated into the front end that shows distance from the user to a
set of closest stores. It also returns a GeoJSON that can be used to chart a path from the user to
the store. While I did show it’s basic functionality, extras such as traffic and turn by turn
directions are unavailable on the free tier and if this feature was to be incorporated fully into the
app, it would require paying for the API.